Updated test files structure
