+++ checkout/tests/test_urls.py
@@ -1,9 +1,8 @@
-from django.test import TestCase, SimpleTestCase
+from django.test import SimpleTestCase
 from django.urls import reverse, resolve
 from checkout.views import checkout
 
 
-# URLS
 class TestCheckoutUrls(SimpleTestCase):
 
     def test_checkout_url_is_resolved(self):
