+++ search/tests/test_urls.py
@@ -1,9 +1,8 @@
-from django.test import TestCase, SimpleTestCase
+from django.test import SimpleTestCase
 from django.urls import reverse, resolve
 from search.views import do_search
 
 
-# URLS
 class TestSearchUrls(SimpleTestCase):
 
     def test_search_url_is_resolved(self):
